+++ controllers/user.ts
@@ -1,10 +1,8 @@
 import { type ICustomRequest, type IUserRequest, type IUserResponse, Status } from '../dto'
 import { type Request, type Response } from 'express'
 import { User } from '../models'
-import { createUserInstanceService, createUserService, getUserbyIdService } from '../services/user'
+import { createFilter, createUserInstanceService, createUserService, errorMessageHandler, getUserbyIdService } from '../services'
 import { userRepository } from '../repositories'
-import createFilter from '../services/createFilter'
-import errorMessageHandler from '../services/errorMessage'
 
 const createUser = async (req: ICustomRequest, res: Response): Promise<Response> => {
   try {
