+++ routes/user.ts
@@ -1,8 +1,8 @@
 import { Roles } from '../dto'
 import { Router } from 'express'
 import { body, param } from 'express-validator'
-import { checkIfRoleIsSent, existsUserById, userValidationRules } from '../services/user'
 import { createUser, deleteUser, getAllUsers, getMoviesLikedByUser, getUserById, getUserProfile, updateUser } from '../controllers'
+import { checkIfRoleIsSent, existsUserById, userValidationRules } from '../services'
 import { validateEmailInChange, validateFields, validateJWT, validateQuery, validateRole, validateRoleOnCreate, validateStatus } from '../middlewares'
 
 const userRouter = Router()
